chore: remove products page component
